requirement.txt
requests
flask
Pillow
scipy
pytorch-pretrained-biggan
nltk